Redesigned the CameraTab
Now it's easier to create camera contexts that fit nicely in a specific area.

+++ Sokoban/src/animation.h
@@ -41,7 +41,7 @@ class GateTransitionAnimation: public Animation {
 public:
     GateTransitionAnimation(bool state);
     ~GateTransitionAnimation();
-    float height();
+    double height();
 
 private:
     bool state_;
