+++ Sokoban/src/roommap.cpp
@@ -312,8 +312,9 @@ void ObjectDrawer::operator()(int id, Point3 pos) {
 	}
 }
 
-void RoomMap::draw(GraphicsManager* gfx, float angle) {
+void RoomMap::draw(GraphicsManager* gfx, double angle) {
 	GameObjIDPosFunc drawer = ObjectDrawer{ obj_array_, gfx };
+	// TODO: use a smarter map rectangle!
 	for (auto& layer : layers_) {
 		layer->apply_to_rect_with_pos(MapRect{ 0,0,width_,height_ }, drawer);
 	}
