+++ Sokoban/src/shader.cpp
@@ -72,8 +72,8 @@ void Shader::setInt(const std::string &name, int value) const {
     glUniform1i(glGetUniformLocation(ID, name.c_str()), value);
 }
 
-void Shader::setFloat(const std::string &name, float value) const {
-    glUniform1f(glGetUniformLocation(ID, name.c_str()), value);
+void Shader::setFloat(const std::string &name, double value) const {
+    glUniform1f(glGetUniformLocation(ID, name.c_str()), (float)value);
 }
 
 void Shader::setMat4(const std::string &name, const glm::mat4 &mat) const {
