+++ Sokoban/src/snakeblock.h
@@ -36,7 +36,7 @@ public:
     bool can_link(SnakeBlock*);
 
     void draw(GraphicsManager*);
-	void draw_force_indicators(GraphicsManager* gfx, FPoint3 p, float radius);
+	void draw_force_indicators(GraphicsManager* gfx, FPoint3 p, double radius);
 
     bool available();
     bool confused(RoomMap*);
